refactor: use generic ZodSchemaValidation, that receives a schema paramter, to remove repeated code
