+++ src/@core/backend/main/factories/controllers/cart/addCartItemControllerFactory.ts
@@ -1,11 +1,12 @@
 import { DbAddCartItemUseCase } from "@/@core/backend/domain/usecases/cart/dbAddCartItemUseCase"
-import { ZodCartItemValidator } from "@/@core/backend/infra/services/validators/cart/zodCartItemValidator"
+import { zodCartItemSchema } from "@/@core/backend/infra/services/validators/cart/zodCartItemValidator"
 import { AddCartItemController } from "@/@core/backend/presentation/controllers/cart/addCartItemController"
 import { mongoProductRepository } from "../../repositories/productRepositoryFactory"
 import { createMongoCartRepository } from "../../repositories/cartRepositoryFactory"
+import { ZodSchemaValidator } from "@/@core/backend/infra/services/zodSchemaValidator"
 
 function createAddCartItemController() {
-    const zodCartItemValidator = new ZodCartItemValidator()
+    const zodCartItemValidator = new ZodSchemaValidator(zodCartItemSchema)
     const mongoCartRepository = createMongoCartRepository()
 
     const dbAddCartItemUseCase = new DbAddCartItemUseCase(
